Camera Movement and Rotation
Added camera movement and smooth rotation about a rig. This will
hopefully make it easy to limit the extends of movement to the map area.

The movement of the rig is based upon input from the standard
horizontal and vertical axis controls (WASD or arrow keys on a
keyboard).

The camera rotation about the rig occurs over a settable time interval
(set to 0.1s).

The rig movement responds to the current angle of the camera so that
the movement controls do not get skewed.